+++ attributes/elasticsearch_exporter.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+default['elasticsearch_exporter']['version'] = '1.1.0'