+++ attributes/filebeat.rb
@@ -1,4 +1,8 @@
 # frozen_string_literal: true
 
-default['filebeat']['version']                             = '7.2.0'
-default['filebeat']['conf_path']                           = '/etc/filebeat'
+default['filebeat']['version']           = '7.4.2'
+default['filebeat']['conf_path']         = '/etc/filebeat'
+default['filebeat']['bulk_max_size']     = '2048'
+default['filebeat']['number_of_shards']  = '1'
+default['filebeat']['logs']['external']  = 'false'
+default['filebeat']['logs']['hostname']  = ''