+++ attributes/heartbeat.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-default['heartbeat']['version']                             = '7.2.0'
+default['heartbeat']['version']                             = '7.4.2'
 default['heartbeat']['conf_path']                           = '/etc/heartbeat'