+++ attributes/nlb.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+default['nlb']['register']              = 'false'
+default['nlb']['target_groups']['9200'] = ''
+default['nlb']['target_groups']['9300'] = ''
+default['nlb']['target_groups']['5601'] = ''