+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+### 1.1.0
+* Replace ELB with NLB
+* Support sending logs to external ES instance through Filebeat
+* Add a safe-shutdown process
+* Add an ES exporter for Prometheus scraping
+* Support custom plugin installation
+
 ### 1.0.0
 * Initial release